LoadGen: Fix singed comparison missmatch warning.
Former-commit-id: c79dab0cd53a31a13071f88bc7b4c79f1536ce69

+++ loadgen/loadgen.cc
@@ -350,7 +350,7 @@ std::vector<QueryMetadata> GenerateQueries(
       size_t num_full_repeats = samples_per_query / num_loaded_samples;
       int remainder = samples_per_query % (num_loaded_samples);
 
-      for (int i = 0; i < num_full_repeats; ++i) {
+      for (size_t i = 0; i < num_full_repeats; ++i) {
         std::copy(loaded_samples.begin(), loaded_samples.end(),
                   samples.begin() + i * num_loaded_samples);
       }
